requirements.txt
requests
pandas
scikit-learn
fastapi
uvicorn
mlflow
dagshub
dvc
openmeteo-requests
requests-cache
retry-requests
prometheus-fastapi-instrumentator
pytest
httpx
ydata-profiling
python-dotenv
joblib
matplotlib
seaborn
